Title: 배열의 특정 위치 찾기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240930/배열의 특정 위치 찾기/find-specific-location-fo-array.c
@@ -20,10 +20,10 @@ int main() {
         }
     }
 
-    float avg = 0;
-    avg = b/count;
+    double avg = 0;
+    avg = (double)b/count;
 
-    printf("%d %.1f", a,avg);
+    printf("%d %.1lf", a,avg);
     
     return 0;
 }
